lnapp: refactor
