+++ ptarmd/Makefile
@@ -45,6 +45,8 @@ endif
 C_SOURCE_FILES += $(PRJ_PATH)/ptarmd.c
 C_SOURCE_FILES += $(PRJ_PATH)/p2p.c
 C_SOURCE_FILES += $(PRJ_PATH)/lnapp.c
+C_SOURCE_FILES += $(PRJ_PATH)/lnapp_cb.c
+C_SOURCE_FILES += $(PRJ_PATH)/lnapp_util.c
 C_SOURCE_FILES += $(PRJ_PATH)/lnapp_manager.c
 C_SOURCE_FILES += $(PRJ_PATH)/cmd_json.c
 C_SOURCE_FILES += $(PRJ_PATH)/monitoring.c
